# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminOrderRead": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "$ref": "#/components/schemas/OrderAddressRead"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "currency": {
            "const": "ILS",
            "title": "Currency",
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/OrderHistoryRead"
            },
            "title": "History",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "payment_deadline": {
            "format": "date-time",
            "title": "Payment Deadline",
            "type": "string"
          },
          "refund": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminRefundRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "shipment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShipmentRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "shipping_agorot": {
            "title": "Shipping Agorot",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/OrderState"
          },
          "subtotal_agorot": {
            "title": "Subtotal Agorot",
            "type": "integer"
          },
          "total_agorot": {
            "title": "Total Agorot",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "order_number",
          "state",
          "items",
          "subtotal_agorot",
          "shipping_agorot",
          "total_agorot",
          "currency",
          "address",
          "payment_deadline",
          "history",
          "shipment",
          "allowed_actions",
          "created_at",
          "refund"
        ],
        "title": "AdminOrderRead",
        "type": "object"
      },
...
      "AdminRefundRead": {
        "properties": {
          "amount_agorot": {
            "title": "Amount Agorot",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RefundState"
          }
        },
        "required": [
          "id",
          "amount_agorot",
          "state",
          "reason"
        ],
        "title": "AdminRefundRead",
        "type": "object"
      },
...
      "CheckoutRead": {
        "additionalProperties": false,
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderRead"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentIntentRead"
          }
        },
        "required": [
          "order",
          "payment"
        ],
        "title": "CheckoutRead",
        "type": "object"
      },
      "CheckoutRequest": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CheckoutAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "address_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Id"
          },
          "expected_shipping_agorot": {
            "minimum": 0.0,
            "title": "Expected Shipping Agorot",
            "type": "integer"
          }
        },
        "required": [
          "expected_shipping_agorot"
        ],
        "title": "CheckoutRequest",
        "type": "object"
      },
...
      "ConfirmedReasonCommand": {
        "additionalProperties": false,
        "properties": {
          "confirm_order_number": {
            "maxLength": 32,
            "minLength": 5,
            "title": "Confirm Order Number",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "confirm_order_number"
        ],
        "title": "ConfirmedReasonCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderAddressRead": {
        "additionalProperties": false,
        "properties": {
          "apartment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apartment"
          },
          "building": {
            "title": "Building",
            "type": "string"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "country": {
            "const": "IL",
            "title": "Country",
            "type": "string"
          },
          "phone_e164": {
            "title": "Phone E164",
            "type": "string"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "recipient_name": {
            "title": "Recipient Name",
            "type": "string"
          },
          "street": {
            "title": "Street",
            "type": "string"
          }
        },
        "required": [
          "recipient_name",
          "phone_e164",
          "street",
          "building",
          "apartment",
          "city",
          "postal_code",
          "country"
        ],
        "title": "OrderAddressRead",
        "type": "object"
      },
      "OrderHistoryRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "from_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderState"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "to_state": {
            "$ref": "#/components/schemas/OrderState"
          }
        },
        "required": [
          "from_state",
          "to_state",
          "source",
          "reason",
          "created_at"
        ],
        "title": "OrderHistoryRead",
        "type": "object"
      },
      "OrderItemRead": {
        "additionalProperties": false,
        "properties": {
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Attributes",
            "type": "object"
          },
          "currency": {
            "const": "ILS",
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "line_total_agorot": {
            "title": "Line Total Agorot",
            "type": "integer"
          },
          "machine_model_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Model Id"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "product_name_he": {
            "title": "Product Name He",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "sku_code": {
            "title": "Sku Code",
            "type": "string"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          },
          "unit_price_agorot": {
            "title": "Unit Price Agorot",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sku_id",
          "product_id",
          "product_name_he",
          "sku_code",
          "attributes",
          "unit_price_agorot",
          "quantity",
          "line_total_agorot",
          "currency",
          "machine_model_id"
        ],
        "title": "OrderItemRead",
        "type": "object"
      },
      "OrderQueueRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "format": "uuid",
            "title": "Customer Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "total_agorot": {
            "title": "Total Agorot",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_number",
          "customer_id",
          "state",
          "total_agorot",
          "created_at",
          "updated_at"
        ],
        "title": "OrderQueueRead",
        "type": "object"
      },
      "OrderRead": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "$ref": "#/components/schemas/OrderAddressRead"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "currency": {
            "const": "ILS",
            "title": "Currency",
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/OrderHistoryRead"
            },
            "title": "History",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "payment_deadline": {
            "format": "date-time",
            "title": "Payment Deadline",
            "type": "string"
          },
          "shipment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShipmentRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "shipping_agorot": {
            "title": "Shipping Agorot",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/OrderState"
          },
          "subtotal_agorot": {
            "title": "Subtotal Agorot",
            "type": "integer"
          },
          "total_agorot": {
            "title": "Total Agorot",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "order_number",
          "state",
          "items",
          "subtotal_agorot",
          "shipping_agorot",
          "total_agorot",
          "currency",
          "address",
          "payment_deadline",
          "history",
          "shipment",
          "allowed_actions",
          "created_at"
        ],
        "title": "OrderRead",
        "type": "object"
      },
      "OrderState": {
        "enum": [
          "pending_payment",
          "paid",
          "processing",
          "shipped",
          "delivered",
          "payment_expired",
          "cancelled",
          "refunded"
        ],
        "title": "OrderState",
        "type": "string"
      },
...
      "RefundRead": {
        "additionalProperties": false,
        "properties": {
          "amount_agorot": {
            "title": "Amount Agorot",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "payment_id": {
            "format": "uuid",
            "title": "Payment Id",
            "type": "string"
          },
          "provider_refund_id": {
            "title": "Provider Refund Id",
            "type": "string"
          },
          "refund_id": {
            "format": "uuid",
            "title": "Refund Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RefundState"
          }
        },
        "required": [
          "refund_id",
          "provider_refund_id",
          "payment_id",
          "amount_agorot",
          "currency",
          "state"
        ],
        "title": "RefundRead",
        "type": "object"
      },
...
      "ShipOrderCommand": {
        "additionalProperties": false,
        "properties": {
          "carrier": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Carrier",
            "type": "string"
          },
          "tracking_number": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Tracking Number",
            "type": "string"
          },
          "tracking_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Url"
          }
        },
        "required": [
          "carrier",
          "tracking_number"
        ],
        "title": "ShipOrderCommand",
        "type": "object"
      },
      "ShipmentRead": {
        "additionalProperties": false,
        "properties": {
          "carrier": {
            "title": "Carrier",
            "type": "string"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "shipped_at": {
            "format": "date-time",
            "title": "Shipped At",
            "type": "string"
          },
          "tracking_number": {
            "title": "Tracking Number",
            "type": "string"
          },
          "tracking_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Url"
          }
        },
        "required": [
          "carrier",
          "tracking_number",
          "tracking_url",
          "shipped_at",
          "delivered_at"
        ],
        "title": "ShipmentRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Coffix API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "list_order_queue_api_v1_admin_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 160,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderQueueRead"
                  },
                  "title": "Response List Order Queue Api V1 Admin Orders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Order Queue",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/orders/{order_id}": {
      "get": {
        "operationId": "get_admin_order_api_v1_admin_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Admin Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/admin/orders/{order_id}/cancel": {
      "post": {
        "operationId": "cancel_order_api_v1_admin_orders__order_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmedReasonCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Order",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/admin/orders/{order_id}/process": {
      "post": {
        "operationId": "process_order_api_v1_admin_orders__order_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/admin/orders/{order_id}/refund": {
      "post": {
        "operationId": "refund_order_api_v1_admin_orders__order_id__refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmedReasonCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refund Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/admin/orders/{order_id}/ship": {
      "post": {
        "operationId": "ship_order_api_v1_admin_orders__order_id__ship_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipOrderCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ship Order",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/checkout": {
      "post": {
        "operationId": "checkout_api_v1_checkout_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Checkout",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
...
  }
}